Overhaul
Added a ton of functionality. Specific cards are identified by those stored in the EEPROM by the setup function. Added functionality to better handle discrepancies during the deal of the cards, ie duplicate cards showing up at player readers. Added a debugging mode that is initiated by making pin A1 high. In this mode, more data is output to the serial monitor. Also added a LED on pin A5 that blinks when a card is successfully read. The regular output to the serial monitor is formatted for easy post-processing for overlay on a video broadcast of the hand.